requirements.txt
open3d==0.18.0
mediapy==1.1.2
lpips==0.1.4
scikit-image==0.21.0
tqdm==4.66.2
trimesh==4.3.2
plyfile
opencv-python
GPUtil
ninja
ffmpeg-python
pillow==10.2.0